Fading LED on Analog Read SUCCES
Changed the code in the Master to successful code. Direct copy of the code in the newly uploaded file.
Names differ, rewrote the program in Arduino IDE under different name (need to start paying attention to my Documental Hygiene).
Changes are also commented on in the new file and the commits regarding.